ordering 2,3,4 OK. Working on ordering 5

+++ srcs/check_arg.c
@@ -6,7 +6,7 @@
 /*   By: tpereira <tpereira@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/25 11:18:14 by tpereira          #+#    #+#             */
-/*   Updated: 2022/01/31 15:48:26 by tpereira         ###   ########.fr       */
+/*   Updated: 2022/02/01 20:02:17 by tpereira         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@ int	check_order(t_stack *a)
 		return (0);
 	while (temp->next != NULL)
 	{
-		if ((int)temp->content > (int)temp->next->content)
+		if (temp->content > temp->next->content)
 			return (0);
 		temp = temp->next;
 	}
